requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
blis==1.3.3
catalogue==2.0.10
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
emoji==2.15.0
en_core_web_sm==3.8.0
fastapi==0.132.0
filelock==3.25.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.3.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
installer==0.7.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
kombu==5.6.2
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.7
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
olefile==0.47
onnx==1.20.1
onnxruntime==1.24.3
openai==2.24.0
opencv-python==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pi_heif==1.3.0
pikepdf==10.3.0
pillow==12.1.1
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.7.5
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.2.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
tiktoken==0.12.0
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
webencodings==0.5.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
